+++ xmake/actions/install/main.lua
@@ -30,6 +30,26 @@ import("core.base.privilege")
 import("privilege.sudo")
 import("install")
 
+-- get install directory
+function _installdir()
+
+    -- the install directory
+    --
+    -- DESTDIR: be compatible with https://www.gnu.org/prep/standards/html_node/DESTDIR.html
+    --
+    local installdir = option.get("installdir") or os.getenv("INSTALLDIR") or os.getenv("DESTDIR") or platform.get("installdir")
+    assert(installdir, "unknown install directory!")
+
+    -- append prefix
+    local prefix = option.get("prefix") or os.getenv("PREFIX")
+    if prefix then
+        installdir = path.join(installdir, prefix)
+    end
+
+    -- ok
+    return installdir
+end
+
 -- main
 function main()
 
@@ -39,8 +59,11 @@ function main()
     -- build it first
     task.run("build", {target = targetname, all = option.get("all")})
 
+    -- get install directory
+    local installdir = _installdir()
+
     -- trace
-    print("installing to %s ...", option.get("installdir") or platform.get("installdir"))
+    print("installing to %s ...", installdir)
 
     -- attempt to install directly
     try
@@ -48,7 +71,7 @@ function main()
         function ()
 
             -- install target
-            install(targetname or ifelse(option.get("all"), "__all", "__def"))
+            install(targetname or ifelse(option.get("all"), "__all", "__def"), installdir)
 
             -- trace
             cprint("${bright}install ok!${clear}${ok_hand}")
@@ -66,7 +89,7 @@ function main()
                         function ()
 
                             -- install target
-                            install(targetname or ifelse(option.get("all"), "__all", "__def"))
+                            install(targetname or ifelse(option.get("all"), "__all", "__def"), installdir)
 
                             -- trace
                             cprint("${bright}install ok!${clear}${ok_hand}")
@@ -99,7 +122,7 @@ function main()
                     if answer == 'y' or answer == '' then
 
                         -- install target with administrator permission
-                        sudo.runl(path.join(os.scriptdir(), "install_admin.lua"), {targetname or ifelse(option.get("all"), "__all", "__def"), option.get("installdir")})
+                        sudo.runl(path.join(os.scriptdir(), "install_admin.lua"), {targetname or ifelse(option.get("all"), "__all", "__def"), installdir})
 
                         -- trace
                         cprint("${bright}install ok!${clear}${ok_hand}")