+++ xmake/actions/install/xmake.lua
@@ -45,11 +45,19 @@ task("install")
                 -- options
             ,   options = 
                 {
-                    {'o', "installdir", "kv", nil,      "Set the install directory."    }
-                ,   {'a', "all",        "k",  nil,      "Install all targets."          }
+                    {'o', "installdir", "kv", nil,      "Set the install directory.",
+                                                        ".e.g",
+                                                        "    $ xmake install -o /usr/local",
+                                                        "or  $ export DESTDIR=/usr/local; xmake install",
+                                                        "or  $ export INSTALLDIR=/usr/local; xmake install" }
+                ,   {'p', "prefix", "kv", nil,          "Set the prefix directory.",
+                                                        ".e.g",
+                                                        "    $ xmake install --prefix=local",
+                                                        "or  $ export PREFIX=local; xmake install"          }
+                ,   {'a', "all",        "k",  nil,      "Install all targets."                              }
 
-                ,   {}
-                ,   {nil, "target",     "v",  nil,      "Install the given target."     }
+                ,   {                                                                                       }
+                ,   {nil, "target",     "v",  nil,      "Install the given target."                         }
                 }
             }