+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Support `@loader_path` and `$ORIGIN` for `add_rpathdirs()`
 * Improve `set_version("x.x.x", {build = "%Y%m%d%H%M"})` and add build version
 * Move docs directory to xmake-docs repo
+* Improve install and uninstall actions and support DESTDIR and PREFIX envirnoment variables
 
 ### Bugs fixed
 
@@ -356,6 +357,7 @@
 * 对`add_rpathdirs()`增加对`@loader_path`和`$ORIGIN`的内置变量支持，提供可迁移动态库加载
 * 改进`set_version("x.x.x", {build = "%Y%m%d%H%M"})` 支持buildversion设置
 * 移除docs目录，将其放置到独立xmake-docs仓库中，减少xmake.zip的大小，优化下载安装的效率
+* 改进安装和卸载脚本，支持DESTDIR和PREFIX环境变量设置
 
 ### Bugs修复